html/browsers/browsing-the-web/navigating-across-documents/009.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [6, 11, 12]
  - id: browsing-the-web.html#javascript-protocol
  - id: form-control-infrastructure.html#concept-form-submit
  - id: form-control-infrastructure.html#submit-get-action
  - id: form-control-infrastructure.html#plan-to-navigate
  - id: form-control-infrastructure.html#planned-navigation

description: |
  1. This test creates a form with `target="test"` and an anchor with both an onclick and href.
  2. The test "clicks" the anchor, which will trigger the onclick handler operation, which will attempt to submit the form. The form's `action` is a `javascript:` url, which has special handling in step 12
  3. The form's `action` will complete, followed by navigating the iframe to the anchor's `href=""`.

/-->
<title>Link with onclick form submit to javascript: url with document.write and href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:document.write('<script>parent.postMessage(&quot;write&quot;, &quot;*&quot;)</script>');"></form>
<a target="test" onclick="document.forms[0].submit()" href="href.html">Test</a>
<script>

var t = async_test();
var messages = [];
t.step(function() {
  document.getElementsByTagName("a")[0].click();
});
onmessage = t.step_func(function(e) {
  messages.push(e.data);
  if (messages.length === 2) {
    assert_array_equals(messages, ["write", "href"]);
    t.done();
  }
});
</script>
